Edição de Cliente
Foi adicionado View, Action e método para Edição de cliente. Fiz uma alteração dentro da classe de modelo do Cliente e quando rodei o comando Update-Database ocorreu um erro. Para solucionar eu deletei o banco de dados e a pasta Migrations e criei novamente.

+++ Benefits/Migrations/20191027002136_CriacaoDoBanco.cs
@@ -12,7 +12,7 @@ namespace Benefits.Migrations
                 name: "Clientes",
                 columns: table => new
                 {
-                    ClienteId = table.Column<int>(nullable: false)
+                    ClienteId = table.Column<long>(nullable: false)
                         .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
                     Nome = table.Column<string>(nullable: true),
                     Email = table.Column<string>(nullable: true),
