+++ Benefits/Models/Cliente.cs
@@ -12,7 +12,7 @@ namespace Benefits.Models
     public class Cliente
     {
         [Key]
-        public int ClienteId { get; set; }
+        public long ClienteId { get; set; }
         public string Nome { get; set; }
         public string Email { get; set; }
         public string Telefone { get; set; }
